Refactor soft constraints to simply the algorithm and fix potential issues. (#520)
* Refactor soft constraints to simply the algorithm and fix potential issues.

1. Check for zero weight so as to avoid unnecessary calculations.
2. Simply the soft constraint interfaces and implementations. Avoid duplicate code.
3. Adjust partition movements constraint logic to reduce the chance of moving partition when the baseline and best possible assignment diverge.
4. Estimate utilization in addition to the other usage estimation. The estimation will be used as a base when calculating the capacity usage score. This is to ensure the algorithm treats different clusters with different overall usage in the same way.
5. Fix the issue that high utilization calculation does not consider the current proposed replica usage.
6. Use Sigmoid to calculate usage-based soft constraints score. This enhances the assignment result of the algorithm.
7. Adjust the related test cases.

+++ helix-core/src/test/java/org/apache/helix/controller/rebalancer/waged/constraints/TestSoftConstraintNormalizeFunction.java
@@ -32,14 +32,14 @@ public class TestSoftConstraintNormalizeFunction {
     int minScore = 0;
     SoftConstraint softConstraint = new SoftConstraint(maxScore, minScore) {
       @Override
-      protected float getAssignmentScore(AssignableNode node, AssignableReplica replica,
+      protected double getAssignmentScore(AssignableNode node, AssignableReplica replica,
           ClusterContext clusterContext) {
         return 0;
       }
     };
 
     for (int i = minScore; i <= maxScore; i++) {
-      float normalized = softConstraint.getNormalizeFunction().scale(i);
+      double normalized = softConstraint.getNormalizeFunction().scale(i);
       Assert.assertTrue(normalized <= 1 && normalized >= 0,
           String.format("input: %s, output: %s", i, normalized));
     }
